+++ src/main/java/traben/entity_model_features/models/anim/EMFParser/MathAction.java
@@ -1,5 +1,7 @@
 package traben.entity_model_features.models.anim.EMFParser;
 
+import net.fabricmc.loader.api.FabricLoader;
+
 public enum MathAction  implements MathComponent{
     add,
     subtract,
@@ -39,7 +41,7 @@ public enum MathAction  implements MathComponent{
     }
 
     @Override
-    public Double get() throws EMFMathException {
+    public Float get() throws EMFMathException {
         throw new EMFMathException("ERROR: math action incorrectly called ["+this+"].");
        // return Double.NaN;
     }
