+++ src/main/java/traben/entity_model_features/models/anim/EMFParser/MathVariableConstant.java
@@ -2,17 +2,17 @@ package traben.entity_model_features.models.anim.EMFParser;
 
 import java.util.function.Supplier;
 
-public class MathVariableConstant extends MathValue implements Supplier<Double> , MathComponent{
+public class MathVariableConstant extends MathValue implements Supplier<Float> , MathComponent{
 
-    Double hardCodedValue;
+    Float hardCodedValue;
 
-    public MathVariableConstant(double number, boolean isNegative){
+    public MathVariableConstant(float number, boolean isNegative){
         super(isNegative, null);
         hardCodedValue = number;
     }
 
     @Override
-    public Supplier<Double> getSupplier() {
+    public Supplier<Float> getSupplier() {
         return ()-> hardCodedValue;
     }
